backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
pyvesync==2.1.16
eventlet==0.39.0
python-dotenv==1.0.1
flask-cors==4.0.0